Extracted RestartFlow,Sink and Source to their own files
Helps me keep overview on what im working on.
