+++ src/core/Akka.Streams/Dsl/RestartFlow.cs
@@ -11,302 +11,7 @@ using Akka.Streams.Stage;
 
 namespace Akka.Streams.Dsl
 {
-    /// <summary>
-    /// A RestartSource wraps a <see cref="Source"/> that gets restarted when it completes or fails.
-    /// They are useful for graphs that need to run for longer than the <see cref="Source"/> can necessarily guarantee it will, for
-    /// example, for <see cref="Source"/> streams that depend on a remote server that may crash or become partitioned. The
-    /// RestartSource ensures that the graph can continue running while the <see cref="Source"/> restarts.
-    /// </summary>
-    public static class RestartSource
-    {
-        /// <summary>
-        /// Wrap the given <see cref="Source"/> with a <see cref="Source"/> that will restart it when it fails or complete using an exponential
-        /// backoff.
-        /// This <see cref="Source"/> will never emit a complete or failure, since the completion or failure of the wrapped <see cref="Source"/>
-        /// is always handled by restarting it. The wrapped <see cref="Source"/> can however be cancelled by cancelling this <see cref="Source"/>.
-        /// When that happens, the wrapped <see cref="Source"/>, if currently running will be cancelled, and it will not be restarted.
-        /// This can be triggered simply by the downstream cancelling, or externally by introducing a <see cref="IKillSwitch"/> right
-        /// after this <see cref="Source"/> in the graph.
-        /// This uses the same exponential backoff algorithm as <see cref="BackoffOptions"/>.
-        /// </summary>
-        /// <param name="sourceFactory">A factory for producing the <see cref="Source"/> to wrap.</param>
-        /// <param name="minBackoff">Minimum (initial) duration until the child actor will started again, if it is terminated</param>
-        /// <param name="maxBackoff">The exponential back-off is capped to this duration</param>
-        /// <param name="randomFactor">After calculation of the exponential back-off an additional random delay based on this factor is added, e.g. `0.2` adds up to `20%` delay. In order to skip this additional delay pass in `0`.</param>
-        [Obsolete("Use the overloaded method which accepts Akka.Stream.RestartSettings instead.")]
-        public static Source<T, NotUsed> WithBackoff<T, TMat>(Func<Source<T, TMat>> sourceFactory, TimeSpan minBackoff, TimeSpan maxBackoff, double randomFactor)
-        {
-            var settings = RestartSettings.Create(minBackoff, maxBackoff, randomFactor);
-            return WithBackoff(sourceFactory, settings);
-        }
-
-        /// <summary>
-        /// Wrap the given <see cref="Source"/> with a <see cref="Source"/> that will restart it when it fails or complete using an exponential
-        /// backoff.
-        /// This <see cref="Source"/> will never emit a complete or failure, since the completion or failure of the wrapped <see cref="Source"/>
-        /// is always handled by restarting it. The wrapped <see cref="Source"/> can however be cancelled by cancelling this <see cref="Source"/>.
-        /// When that happens, the wrapped <see cref="Source"/>, if currently running will be cancelled, and it will not be restarted.
-        /// This can be triggered simply by the downstream cancelling, or externally by introducing a <see cref="IKillSwitch"/> right
-        /// after this <see cref="Source"/> in the graph.
-        /// This uses the same exponential backoff algorithm as <see cref="BackoffOptions"/>.
-        /// </summary>
-        /// <param name="sourceFactory">A factory for producing the <see cref="Source"/> to wrap.</param>
-        /// <param name="minBackoff">Minimum (initial) duration until the child actor will started again, if it is terminated</param>
-        /// <param name="maxBackoff">The exponential back-off is capped to this duration</param>
-        /// <param name="randomFactor">After calculation of the exponential back-off an additional random delay based on this factor is added, e.g. `0.2` adds up to `20%` delay. In order to skip this additional delay pass in `0`.</param>
-        /// <param name="maxRestarts">The amount of restarts is capped to this amount within a time frame of minBackoff. Passing `0` will cause no restarts and a negative number will not cap the amount of restarts.</param>
-        [Obsolete("Use the overloaded method which accepts Akka.Stream.RestartSettings instead.")]
-        public static Source<T, NotUsed> WithBackoff<T, TMat>(Func<Source<T, TMat>> sourceFactory, TimeSpan minBackoff, TimeSpan maxBackoff, double randomFactor, int maxRestarts)
-        {
-            var settings = RestartSettings.Create(minBackoff, maxBackoff, randomFactor).WithMaxRestarts(maxRestarts, minBackoff);
-            return WithBackoff(sourceFactory, settings);
-        }
-
-        /// <summary>
-        /// Wrap the given <see cref="Source"/> with a <see cref="Source"/> that will restart it when it fails or complete using an exponential
-        /// backoff.
-        /// <para>
-        /// This <see cref="Source"/> will never emit a complete or failure, since the completion or failure of the wrapped <see cref="Source"/>
-        /// is always handled by restarting it. The wrapped <see cref="Source"/> can however be cancelled by cancelling this <see cref="Source"/>.
-        /// When that happens, the wrapped <see cref="Source"/>, if currently running will be cancelled, and it will not be restarted.
-        /// This can be triggered simply by the downstream cancelling, or externally by introducing a <see cref="IKillSwitch"/> right
-        /// after this <see cref="Source"/> in the graph.
-        /// </para>
-        /// <para>This uses the same exponential backoff algorithm as <see cref="BackoffOptions"/>.</para>
-        /// </summary>
-        /// <param name="sourceFactory">A factory for producing the <see cref="Source"/> to wrap.</param>
-        /// <param name="settings"><see cref="RestartSettings" /> defining restart configuration</param>
-        public static Source<T, NotUsed> WithBackoff<T, TMat>(Func<Source<T, TMat>> sourceFactory, RestartSettings settings)
-            => Source.FromGraph(new RestartWithBackoffSource<T, TMat>(sourceFactory, settings, onlyOnFailures: false));
-
-        /// <summary>
-        /// Wrap the given <see cref="Source"/> with a <see cref="Source"/> that will restart it when it fails using an exponential backoff.
-        /// This <see cref="Source"/> will never emit a failure, since the failure of the wrapped <see cref="Source"/> is always handled by
-        /// restarting. The wrapped <see cref="Source"/> can be cancelled by cancelling this <see cref="Source"/>.
-        /// When that happens, the wrapped <see cref="Source"/>, if currently running will be cancelled, and it will not be restarted.
-        /// This can be triggered simply by the downstream cancelling, or externally by introducing a <see cref="IKillSwitch"/> right
-        /// after this <see cref="Source"/> in the graph.
-        /// This uses the same exponential backoff algorithm as <see cref="BackoffOptions"/>.
-        /// </summary>
-        /// <param name="sourceFactory">A factory for producing the <see cref="Source"/> to wrap.</param>
-        /// <param name="minBackoff">Minimum (initial) duration until the child actor will started again, if it is terminated</param>
-        /// <param name="maxBackoff">The exponential back-off is capped to this duration</param>
-        /// <param name="randomFactor">After calculation of the exponential back-off an additional random delay based on this factor is added, e.g. `0.2` adds up to `20%` delay. In order to skip this additional delay pass in `0`.</param>
-        [Obsolete("Use the overloaded method which accepts Akka.Stream.RestartSettings instead.")]
-        public static Source<T, NotUsed> OnFailuresWithBackoff<T, TMat>(Func<Source<T, TMat>> sourceFactory, TimeSpan minBackoff, TimeSpan maxBackoff, double randomFactor)
-        {
-            var settings = RestartSettings.Create(minBackoff, maxBackoff, randomFactor);
-            return OnFailuresWithBackoff(sourceFactory, settings);
-        }
-
-        /// <summary>
-        /// Wrap the given <see cref="Source"/> with a <see cref="Source"/> that will restart it when it fails using an exponential backoff.
-        /// This <see cref="Source"/> will never emit a failure, since the failure of the wrapped <see cref="Source"/> is always handled by
-        /// restarting. The wrapped <see cref="Source"/> can be cancelled by cancelling this <see cref="Source"/>.
-        /// When that happens, the wrapped <see cref="Source"/>, if currently running will be cancelled, and it will not be restarted.
-        /// This can be triggered simply by the downstream cancelling, or externally by introducing a <see cref="IKillSwitch"/> right
-        /// after this <see cref="Source"/> in the graph.
-        /// This uses the same exponential backoff algorithm as <see cref="BackoffOptions"/>.
-        /// </summary>
-        /// <param name="sourceFactory">A factory for producing the <see cref="Source"/> to wrap.</param>
-        /// <param name="minBackoff">Minimum (initial) duration until the child actor will started again, if it is terminated</param>
-        /// <param name="maxBackoff">The exponential back-off is capped to this duration</param>
-        /// <param name="randomFactor">After calculation of the exponential back-off an additional random delay based on this factor is added, e.g. `0.2` adds up to `20%` delay. In order to skip this additional delay pass in `0`.</param>
-        /// <param name="maxRestarts">The amount of restarts is capped to this amount within a time frame of minBackoff. Passing `0` will cause no restarts and a negative number will not cap the amount of restarts.</param>
-        [Obsolete("Use the overloaded method which accepts Akka.Stream.RestartSettings instead.")]
-        public static Source<T, NotUsed> OnFailuresWithBackoff<T, TMat>(Func<Source<T, TMat>> sourceFactory, TimeSpan minBackoff, TimeSpan maxBackoff, double randomFactor, int maxRestarts)
-        {
-            var settings = RestartSettings.Create(minBackoff, maxBackoff, randomFactor).WithMaxRestarts(maxRestarts, minBackoff);
-            return OnFailuresWithBackoff(sourceFactory, settings);
-        }
-
-        /// <summary>
-        /// Wrap the given <see cref="Source"/> with a <see cref="Source"/> that will restart it when it fails using an exponential backoff.
-        /// <para>
-        /// This <see cref="Source"/> will never emit a failure, since the failure of the wrapped <see cref="Source"/> is always handled by
-        /// restarting. The wrapped <see cref="Source"/> can be cancelled by cancelling this <see cref="Source"/>.
-        /// When that happens, the wrapped <see cref="Source"/>, if currently running will be cancelled, and it will not be restarted.
-        /// This can be triggered simply by the downstream cancelling, or externally by introducing a <see cref="IKillSwitch"/> right
-        /// after this <see cref="Source"/> in the graph.
-        /// </para>
-        /// <para>This uses the same exponential backoff algorithm as <see cref="BackoffOptions"/>.</para>
-        /// </summary>
-        /// <param name="sourceFactory">A factory for producing the <see cref="Source"/> to wrap.</param>
-        /// <param name="settings"><see cref="RestartSettings" /> defining restart configuration</param>
-        public static Source<T, NotUsed> OnFailuresWithBackoff<T, TMat>(Func<Source<T, TMat>> sourceFactory, RestartSettings settings)
-            => Source.FromGraph(new RestartWithBackoffSource<T, TMat>(sourceFactory, settings, onlyOnFailures: true));
-    }
-
-    internal sealed class RestartWithBackoffSource<T, TMat> : GraphStage<SourceShape<T>>
-    {
-        public Func<Source<T, TMat>> SourceFactory { get; }
-        public RestartSettings Settings { get; }
-        public bool OnlyOnFailures { get; }
-
-        public RestartWithBackoffSource(Func<Source<T, TMat>> sourceFactory, RestartSettings settings, bool onlyOnFailures)
-        {
-            SourceFactory = sourceFactory;
-            Settings = settings;
-            OnlyOnFailures = onlyOnFailures;
-            Shape = new SourceShape<T>(Out);
-        }
-
-        public Outlet<T> Out { get; } = new Outlet<T>("RestartWithBackoffSource.out");
-
-        public override SourceShape<T> Shape { get; }
-
-        protected override GraphStageLogic CreateLogic(Attributes inheritedAttributes) => new Logic(this, "Source");
-
-        private sealed class Logic : RestartWithBackoffLogic<SourceShape<T>, T, T>
-        {
-            private readonly RestartWithBackoffSource<T, TMat> _stage;
-
-            public Logic(RestartWithBackoffSource<T, TMat> stage, string name)
-                : base(name, stage.Shape, null, stage.Out, stage.Settings, stage.OnlyOnFailures)
-            {
-                _stage = stage;
-                Backoff();
-            }
-
-            protected override void StartGraph()
-            {
-                var sinkIn = CreateSubInlet(_stage.Out);
-                _stage.SourceFactory().RunWith(sinkIn.Sink, SubFusingMaterializer);
-                if (IsAvailable(_stage.Out))
-                    sinkIn.Pull();
-            }
-
-            protected override void Backoff() => SetHandler(_stage.Out, () =>
-            {
-                // do nothing
-            });
-        }
-    }
-
-    /// <summary>
-    /// A RestartSink wraps a <see cref="Sink"/> that gets restarted when it completes or fails.
-    /// They are useful for graphs that need to run for longer than the <see cref="Sink"/> can necessarily guarantee it will, for
-    /// example, for <see cref="Sink"/> streams that depend on a remote server that may crash or become partitioned. The
-    /// RestartSink ensures that the graph can continue running while the <see cref="Sink"/> restarts.
-    /// </summary>
-    public static class RestartSink
-    {
-        /// <summary>
-        /// Wrap the given <see cref="Sink"/> with a <see cref="Sink"/> that will restart it when it fails or complete using an exponential
-        /// backoff.
-        /// <para>
-        /// This <see cref="Sink"/> will never cancel, since cancellation by the wrapped <see cref="Sink"/> is always handled by restarting it.
-        /// The wrapped <see cref="Sink"/> can however be completed by feeding a completion or error into this <see cref="Sink"/>. When that
-        /// happens, the <see cref="Sink"/>, if currently running, will terminate and will not be restarted. This can be triggered
-        /// simply by the upstream completing, or externally by introducing a <see cref="IKillSwitch"/> right before this <see cref="Sink"/> in the
-        /// graph.
-        /// The restart process is inherently lossy, since there is no coordination between cancelling and the sending of
-        /// messages. When the wrapped <see cref="Sink"/> does cancel, this <see cref="Sink"/> will backpressure, however any elements already
-        /// sent may have been lost.
-        /// </para>
-        /// <para>This uses the same exponential backoff algorithm as <see cref="BackoffOptions"/>.</para>
-        /// </summary>
-        /// <param name="sinkFactory">A factory for producing the <see cref="Sink"/> to wrap.</param>
-        /// <param name="minBackoff">Minimum (initial) duration until the child actor will started again, if it is terminated</param>
-        /// <param name="maxBackoff">The exponential back-off is capped to this duration</param>
-        /// <param name="randomFactor">After calculation of the exponential back-off an additional random delay based on this factor is added, e.g. `0.2` adds up to `20%` delay. In order to skip this additional delay pass in `0`.</param>
-        [Obsolete("Use the overloaded method which accepts Akka.Stream.RestartSettings instead.")]
-        public static Sink<T, NotUsed> WithBackoff<T, TMat>(Func<Sink<T, TMat>> sinkFactory, TimeSpan minBackoff, TimeSpan maxBackoff, double randomFactor)
-        {
-            var settings = RestartSettings.Create(minBackoff, maxBackoff, randomFactor);
-            return WithBackoff(sinkFactory, settings);
-        }
-
-        /// <summary>
-        /// Wrap the given <see cref="Sink"/> with a <see cref="Sink"/> that will restart it when it fails or complete using an exponential
-        /// backoff.
-        /// <para>
-        /// This <see cref="Sink"/> will never cancel, since cancellation by the wrapped <see cref="Sink"/> is always handled by restarting it.
-        /// The wrapped <see cref="Sink"/> can however be completed by feeding a completion or error into this <see cref="Sink"/>. When that
-        /// happens, the <see cref="Sink"/>, if currently running, will terminate and will not be restarted. This can be triggered
-        /// simply by the upstream completing, or externally by introducing a <see cref="IKillSwitch"/> right before this <see cref="Sink"/> in the
-        /// graph.
-        /// The restart process is inherently lossy, since there is no coordination between cancelling and the sending of
-        /// messages. When the wrapped <see cref="Sink"/> does cancel, this <see cref="Sink"/> will backpressure, however any elements already
-        /// sent may have been lost.
-        /// </para>
-        /// <para>This uses the same exponential backoff algorithm as <see cref="BackoffOptions"/>.</para>
-        /// </summary>
-        /// <param name="sinkFactory">A factory for producing the <see cref="Sink"/> to wrap.</param>
-        /// <param name="minBackoff">Minimum (initial) duration until the child actor will started again, if it is terminated</param>
-        /// <param name="maxBackoff">The exponential back-off is capped to this duration</param>
-        /// <param name="randomFactor">After calculation of the exponential back-off an additional random delay based on this factor is added, e.g. `0.2` adds up to `20%` delay. In order to skip this additional delay pass in `0`.</param>
-        /// <param name="maxRestarts">The amount of restarts is capped to this amount within a time frame of minBackoff. Passing `0` will cause no restarts and a negative number will not cap the amount of restarts.</param>
-        [Obsolete("Use the overloaded method which accepts Akka.Stream.RestartSettings instead.")]
-        public static Sink<T, NotUsed> WithBackoff<T, TMat>(Func<Sink<T, TMat>> sinkFactory, TimeSpan minBackoff, TimeSpan maxBackoff, double randomFactor, int maxRestarts)
-        {
-            var settings = RestartSettings.Create(minBackoff, maxBackoff, randomFactor).WithMaxRestarts(maxRestarts, minBackoff);
-            return WithBackoff(sinkFactory, settings);
-        }
-
-        /// <summary>
-        /// Wrap the given <see cref="Sink"/> with a <see cref="Sink"/> that will restart it when it fails or complete using an exponential
-        /// backoff.
-        /// <para>
-        /// This <see cref="Sink"/> will never cancel, since cancellation by the wrapped <see cref="Sink"/> is always handled by restarting it.
-        /// The wrapped <see cref="Sink"/> can however be completed by feeding a completion or error into this <see cref="Sink"/>. When that
-        /// happens, the <see cref="Sink"/>, if currently running, will terminate and will not be restarted. This can be triggered
-        /// simply by the upstream completing, or externally by introducing a <see cref="IKillSwitch"/> right before this <see cref="Sink"/> in the
-        /// graph.
-        /// The restart process is inherently lossy, since there is no coordination between cancelling and the sending of
-        /// messages. When the wrapped <see cref="Sink"/> does cancel, this <see cref="Sink"/> will backpressure, however any elements already
-        /// sent may have been lost.
-        /// </para>
-        /// <para>This uses the same exponential backoff algorithm as <see cref="BackoffOptions"/>.</para>
-        /// </summary>
-        /// <param name="sinkFactory">A factory for producing the <see cref="Sink"/> to wrap.</param>
-        /// <param name="settings"><see cref="RestartSettings" /> defining restart configuration</param>
-        public static Sink<T, NotUsed> WithBackoff<T, TMat>(Func<Sink<T, TMat>> sinkFactory, RestartSettings settings)
-            => Sink.FromGraph(new RestartWithBackoffSink<T, TMat>(sinkFactory, settings));
-    }
-
-    internal sealed class RestartWithBackoffSink<T, TMat> : GraphStage<SinkShape<T>>
-    {
-        public Func<Sink<T, TMat>> SinkFactory { get; }
-        public RestartSettings Settings { get; }
-
-        public RestartWithBackoffSink(Func<Sink<T, TMat>> sinkFactory, RestartSettings settings)
-        {
-            SinkFactory = sinkFactory;
-            Settings = settings;
-            Shape = new SinkShape<T>(In);
-        }
-
-        public Inlet<T> In { get; } = new Inlet<T>("RestartWithBackoffSink.in");
-
-        public override SinkShape<T> Shape { get; }
-
-        protected override GraphStageLogic CreateLogic(Attributes inheritedAttributes) => new Logic(this, "Sink");
-
-        private sealed class Logic : RestartWithBackoffLogic<SinkShape<T>, T, T>
-        {
-            private readonly RestartWithBackoffSink<T, TMat> _stage;
-
-            public Logic(RestartWithBackoffSink<T, TMat> stage, string name)
-                : base(name, stage.Shape, stage.In, null, stage.Settings, onlyOnFailures: false)
-            {
-                _stage = stage;
-                Backoff();
-            }
-
-            protected override void StartGraph()
-            {
-                var sourceOut = CreateSubOutlet(_stage.In);
-                Source.FromGraph(sourceOut.Source).RunWith(_stage.SinkFactory(), SubFusingMaterializer);
-            }
-
-            protected override void Backoff() => SetHandler(_stage.In, () =>
-            {
-                // do nothing
-            });
-        }
-    }
-
+  
     /// <summary>
     /// A RestartFlow wraps a <see cref="Flow"/> that gets restarted when it completes or fails.
     /// They are useful for graphs that need to run for longer than the <see cref="Flow"/> can necessarily guarantee it will, for
@@ -499,7 +204,10 @@ namespace Akka.Streams.Dsl
             {
                 var sourceOut = CreateSubOutlet(_stage.In);
                 var sinkIn = CreateSubInlet(_stage.Out);
-                Source.FromGraph(sourceOut.Source).Via(_stage.FlowFactory()).RunWith(sinkIn.Sink, SubFusingMaterializer);
+                
+                Source.FromGraph(sourceOut.Source)
+                    .Via(_stage.FlowFactory())
+                    .RunWith(sinkIn.Sink, SubFusingMaterializer);
                 if (IsAvailable(_stage.Out))
                     sinkIn.Pull();
 
@@ -587,10 +295,12 @@ namespace Akka.Streams.Dsl
                         Complete(Out);
                     else
                     {
-                        Log.Debug("Restarting graph due to finished upstream");
                         ScheduleRestartTimer();
                     }
                 },
+                /*
+                 * upstream in this context is the wrapped stage
+                 */
                 onUpstreamFailure: ex =>
                 {
                     if (_finishing || MaxRestartsReached())
@@ -633,7 +343,6 @@ namespace Akka.Streams.Dsl
                         Cancel(In);
                     else
                     {
-                        Log.Debug("Graph in finished");
                         ScheduleRestartTimer();
                     }
                 }
